Linux: update makelibs script Also removed "echo $scriptdir" from build_app.sh

+++ tools/linux/libs/makelibs.sh
@@ -1,4 +1,12 @@
 #!/bin/sh
+
+olddir=$PWD
+scriptdir=$(dirname "$(readlink -f -- "$0")")
+if [ -z "$scriptdir" ]; then
+	echo Unable to get script dir, exiting! 
+	exit
+fi
+cd $scriptdir
 rm -rf ../obj
 mkdir -p ../obj
 ./makelibs_zlib.sh
@@ -9,3 +17,4 @@ mkdir -p ../obj
 ./makelibs_png.sh
 ./makelibs_squirrel.sh
 ./makelibs_jpeg.sh
+cd $olddir
